+++ panel/models/icon.ts
@@ -138,134 +138,3 @@ export class ClickableIcon extends Control {
     }));
   }
 }
-
-// ToggleIcon
-
-export class ToggleIconView extends ClickableIconView {
-  model: ToggleIcon;
-
-  public *controls() { }
-
-  toggle(): void {
-    if (this.model.disabled) {
-      return;
-    }
-    this.model.value = !this.model.value;
-    this.update_icon()
-  }
-
-  get_icon(): string {
-    return this.model.value ? this.get_active_icon() : this.model.icon;
-  }
-
-  connect_signals(): void {
-    super.connect_signals();
-    const { value } = this.model.properties;
-    this.on_change(value, () => this.update_icon());
-  }
-}
-
-export namespace ToggleIcon {
-  export type Attrs = p.AttrsOf<Props>;
-  export type Props = ClickableIcon.Props & {
-    value: p.Property<boolean>;
-  };
-}
-
-export interface ToggleIcon extends ToggleIcon.Attrs { }
-
-export class ToggleIcon extends ClickableIcon {
-  properties: ToggleIcon.Props;
-  declare __view_type__: ToggleIconView
-  static __module__ = "panel.models.icon";
-
-  constructor(attrs?: Partial<ToggleIcon.Attrs>) {
-    super(attrs);
-  }
-
-  static {
-    this.prototype.default_view = ToggleIconView;
-
-    this.define<ToggleIcon.Props>(({ Boolean }) => ({
-      value: [Boolean, false],
-    }));
-  }
-}
-
-// ButtonIcon
-
-export class ButtonIconView extends ClickableIconView {
-  model: ButtonIcon;
-  _value: boolean;
-  _click_listener: any
-
-  public *controls() { }
-
-  toggle(): void {
-    if (this.model.disabled) {
-      return;
-    }
-    const updateState = (value: boolean, disabled: boolean) => {
-      this._value = value;
-      this.model.disabled = disabled;
-      this.update_icon();
-    };
-    updateState(true, true);
-
-    new Promise(resolve => setTimeout(resolve, this.model.active_duration))
-      .then(() => {
-        updateState(false, false);
-      });
-  }
-
-  update_cursor(): void {
-    this.icon_view.el.style.cursor = this.model.disabled ? 'default' : 'pointer';
-  }
-
-  get_icon(): string {
-    return this._value ? this.get_active_icon() : this.model.icon;
-  }
-
-  render(): void {
-    super.render();
-    this._click_listener = this.increment_clicks.bind(this)
-    this.icon_view.el.addEventListener("click", this._click_listener)
-  }
-
-  increment_clicks() : void {
-    this.model.clicks = this.model.clicks + 1
-  }
-
-  connect_signals(): void {
-    super.connect_signals();
-  }
-}
-
-export namespace ButtonIcon {
-  export type Attrs = p.AttrsOf<Props>;
-  export type Props = ClickableIcon.Props & {
-    clicks: p.Property<number>;
-    active_duration: p.Property<number>;
-  };
-}
-
-export interface ButtonIcon extends ButtonIcon.Attrs { }
-
-export class ButtonIcon extends ClickableIcon {
-  properties: ButtonIcon.Props;
-  declare __view_type__: ButtonIconView
-  static __module__ = "panel.models.icon";
-
-  constructor(attrs?: Partial<ButtonIcon.Attrs>) {
-    super(attrs);
-  }
-
-  static {
-    this.prototype.default_view = ButtonIconView;
-
-    this.define<ButtonIcon.Props>(({ Int }) => ({
-      clicks: [Int, 0],
-      active_duration: [Int, 75],
-    }));
-  }
-}
