+++ panel/models/index.ts
@@ -2,7 +2,7 @@ export {AcePlot} from "./ace"
 export {Audio} from "./audio"
 export {BrowserInfo} from "./browser"
 export {Button} from "./button"
-export {ButtonIcon} from "./icon"
+export {ButtonIcon} from "./button_icon"
 export {Card} from "./card"
 export {CheckboxButtonGroup} from "./checkbox_button_group"
 export {Column} from "./column"
@@ -35,7 +35,7 @@ export {Tabs} from "./tabs"
 export {Terminal} from "./terminal"
 export {TextAreaInput} from "./textarea_input"
 export {TextToSpeech} from "./text_to_speech"
-export {ToggleIcon} from "./icon"
+export {ToggleIcon} from "./toggle_icon"
 export {TooltipIcon} from "./tooltip_icon"
 export {TrendIndicator} from "./trend"
 export {VegaPlot} from "./vega"
